Refactor to separate service
